+++ src/game/tests.rs
@@ -355,6 +355,7 @@ mod vintage_cube_exhume;
 mod vintage_cube_expedition_map;
 mod vintage_cube_exploration;
 mod vintage_cube_faerie_mastermind;
+mod vintage_cube_fanatic_of_rhonas;
 mod vintage_cube_fatal_push;
 mod vintage_cube_fear_of_missing_out;
 mod vintage_cube_fiery_confluence;